move models into model package, tide screens into tide package
