+++ magellan-sample-advanced/src/main/java/com/wealthfront/magellan/sample/advanced/tide/TideLocationsScreen.java
@@ -1,11 +1,12 @@
-package com.wealthfront.magellan.sample.advanced;
+package com.wealthfront.magellan.sample.advanced.tide;
 
 import android.content.Context;
 import android.support.annotation.DrawableRes;
 
 import com.wealthfront.magellan.Screen;
+import com.wealthfront.magellan.sample.advanced.R;
 
-class TideLocationsScreen extends Screen<TideLocationsView> {
+public class TideLocationsScreen extends Screen<TideLocationsView> {
 
   public enum TideLocations {
     SAN_FRANCISCO(9414290, R.drawable.san_francisco, "San Francisco"),
