+++ magellan-sample-advanced/src/test/java/com/wealthfront/magellan/sample/advanced/ExampleUnitTest.java
@@ -15,5 +15,5 @@ public class ExampleUnitTest {
   public void addition_isCorrect() throws Exception {
     assertEquals(4, 2 + 2);
   }
-  
+
 }
